Remove unnecessary components (#3)
