--- client/src/Components/Editor.jsx
@@ -1,29 +1,0 @@
-// import React, { Component } from 'react';
-// import AceEditor from 'react-ace';
-// import 'brace/mode/javascript';
-// import 'brace/theme/monokai';
-// import 'brace/ext/language_tools';
-
-// const Editor = (props) => (
-//   <div>
-//     <AceEditor
-//       mode="javascript"
-//       theme="monokai"
-//       name="ace-editor"
-//       onChange={props.onCodeUpdate}
-//       fontSize={14}
-//       showPrintMargin={true}
-//       showGutter={true}
-//       highlightActiveLine={true}
-//       value={props.code}
-//       enableBasicAutocompletion={true}
-//       enableLiveAutocompletion={true}
-//       enableSnippets={false}
-//       showLineNumbers={true}
-//       tabSize={2}
-//       editorProps={{ $blockScrolling: 'Infinity' }}
-//     />
-//   </div>
-// );
-
-// export default Editor;
